+++ SP4/Base/Source/AI/ProbabilitySystem.cpp
@@ -22,7 +22,7 @@ void CProbabilitySystem::AddProbability(float percentage)
 // Get total number of probabilities
 float CProbabilitySystem::GetTotalNumOfProbabilities()
 {
-	return probabilities.size();
+	return (float)probabilities.size();
 }
 
 // Get a int random probability
@@ -30,24 +30,24 @@ int CProbabilitySystem::GetARandIntProbability()
 {
 	// Maximun percentage of all probabilities
 	int MaxPercentage = 0;
-	for (int i = 0; i < probabilities.size(); i++)
+	for (int i = 0; i < (int)probabilities.size(); i++)
 	{
 		// Add probility's percentage to MaxPercentage
-		MaxPercentage += probabilities[i];
+		MaxPercentage += (int)probabilities[i];
 	}
 
 	// Get a random number within the total percentages
 	int number = Math::RandIntMinMax(0, MaxPercentage);
 
 	// Check which probability the number is in
-	for (int i = 0; i < probabilities.size(); i++)
+	for (int i = 0; i < (int)probabilities.size(); i++)
 	{
 		// if number is lower or equal to the probility's percentage, return the i
-		if (number <= probabilities[i])
+		if (number <= (int)probabilities[i])
 			return i;
 		// else, subtract the probility's percentage off the number
 		else
-			number -= probabilities[i];
+			number -= (int)probabilities[i];
 	}
 	// Return -1, if error
 	return -1;
@@ -58,7 +58,7 @@ int CProbabilitySystem::GetARandFloatProbability()
 {
 	// Maximun percentage of all probabilities
 	float MaxPercentage = 0.0f;
-	for (int i = 0; i < probabilities.size(); i++)
+	for (int i = 0; i < (int)probabilities.size(); i++)
 	{
 		// Add probility's percentage to MaxPercentage
 		MaxPercentage += probabilities[i];
@@ -68,7 +68,7 @@ int CProbabilitySystem::GetARandFloatProbability()
 	float number = Math::RandFloatMinMax(0.0f, MaxPercentage);
 
 	// Check which probability the number is in
-	for (int i = 0; i < probabilities.size(); i++)
+	for (int i = 0; i < (int)probabilities.size(); i++)
 	{
 		// if number is lower or equal to the probility's percentage, return the i
 		if (number <= probabilities[i])
